move save/load indicators to own folder
