+++ src/renderer/src/Game.jsx
@@ -1,7 +1,7 @@
 import { GameCommands } from './components/inputs/GameCommands'
 import { GameFunctions } from './components/funcs/GameFunctions'
-import { SaveIndicator } from './components/SaveIndicator'
-import { LoadIndicator } from './components/LoadIndicator'
+import { SaveIndicator } from './components/ui/indicators/SaveIndicator'
+import { LoadIndicator } from './components/ui/indicators/LoadIndicator'
 
 const Game = () => {
   return (
